+++ packages/next-sitemap/src/utils/file.ts
